+++ netbox_osp/models/__init__.py
@@ -2,6 +2,7 @@ from .cables import OspCable, Tube, Strand
 from .splices import SpliceClosure, SpliceTray, Splice
 from .links import FibreLink, FibreLinkStrand
 from .locations import LocationGeo
+from .trunks import FibreTrunk
 
 __all__ = [
     "OspCable",
@@ -13,4 +14,5 @@ __all__ = [
     "FibreLink",
     "FibreLinkStrand",
     "LocationGeo",
+    "FibreTrunk",
 ]